Improve documentation
Refactor the documentation, moving informations in Tutorial and Reference files:
as some developers will work on the integration of markdown in their nerdz template,
they should be able to work faster if documentation is provided.
It's a boring work, but it could save lives (and time).

Add a link to the Tutorial in mkd.html